refactor console panel
