+++ package.json
@@ -65,6 +65,7 @@
     "react-transition-group": "^4.4.5",
     "rehype-parse": "^8.0.4",
     "rehype-stringify": "^9.0.3",
+    "shiki": "^0.14.3",
     "source-map": "^0.7.4",
     "unified": "^10.1.2",
     "unist-util-visit": "^4.1.2",
