+++ src/assets/locales/en.json
@@ -74,7 +74,7 @@
     "run": "Run",
     "placeholder": "Executable code",
     "newContent": "New content",
-    "error-drawer": {
+    "error-trace": {
       "title": "Error stack detail",
       "fetch-minify-fail": "Fail to fetch origin file",
       "none-sourcemap": "No sourcemap file founded",
