+++ src/assets/locales/zh.json
@@ -74,7 +74,7 @@
     "run": "执行",
     "placeholder": "可执行代码",
     "newContent": "新消息",
-    "error-drawer": {
+    "error-trace": {
       "title": "错误详情",
       "fetch-minify-fail": "获取源文件失败",
       "none-sourcemap": "未找到 SourceMap",
