+++ src/pages/Devtools/ConsolePanel/components/ConsoleNode/ErrorTrace/index.less
